+++ src/collections/contents/contents.js
@@ -51,10 +51,6 @@ Editor.collections.Contents = {
         _id: {
             type: String,
             optional: true
-        },
-        xxxxxx: {   // unused key to be sure we always have something to unset
-            type: String,
-            optional: true
         }
     })
 };