+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Replace all fonts by a converted (and hopefully sanitized) version from https://transfonter.org
     - teSerializer now provides a default collection name
     - Change the document parameter in teSerializer from name to document (bumpring candidate version number)
+    - Change the default edition mode in teSerializer from PREVIEW to STANDARD
     - Publish the accessed document rather that using a method to take advantage of Meteor reactivity
  
 ### 1.3.0