+++ ChangeLog.md
@@ -12,6 +12,7 @@
     - Change the default edition mode in teSerializer from PREVIEW to STANDARD
     - Publish the accessed document rather that using a method to take advantage of Meteor reactivity
     - Uses dburles:mongo-collection-instances package to not try to redeclare existing Mongo collections
+    - Prevent a form to be submitted when hitting Enter during edition
  
 ### 1.3.0